init: basic project settings
