+++ .prettierrc
@@ -1,0 +1,8 @@
+{
+  "max-len": ["error",140,2],
+  "printWidth": 120,
+  "tabWidth": 2,
+  "semi": true,
+  "singleQuote": false,
+  "trailingComma": "all"
+}
