+++ nodemon.json
@@ -1,0 +1,10 @@
+{
+  "ignore": [
+    "node_modules"
+  ],
+  "watch": [
+    "./index.ts"
+  ],
+  "ext": "ts",
+  "exec": "npm start"
+}
